requirements.txt
hazelcast-python-client
Flask
jsonschema